move: constant (columns)
