+++ static_report/src/components/ComparisonReport/CRTabSchemaDetails.tsx
@@ -21,7 +21,7 @@ import {
   zReport,
   ZTableSchema,
 } from '../../types';
-import { NO_VALUE } from '../shared/Columns/ColumnCard/ColumnTypeDetail/constants';
+import { NO_VALUE } from '../shared/Columns/constants';
 import { useLocation } from 'wouter';
 
 type EnrichedColumnData = {
