+++ static_report/src/components/shared/Columns/ColumnCard/ColumnTypeDetail/ColumnTypeDetailCategorical.tsx
@@ -9,7 +9,7 @@ import { getColumnMetricRatio } from '../../../../../utils/transformers';
 import { DataCompositionMatrix } from '../../ColumnMatrices/DataCompositionMatrix';
 import { UniquenessMetrics } from '../../ColumnMatrices/UniquenessMatrix';
 import { MetricCell } from '../../ColumnMetrics/MetricCell';
-import { MODE, VALIDS } from './constants';
+import { MODE, VALIDS } from '../../constants';
 
 type Props = { columnDatum: ColumnSchema };
 export const ColumnTypeDetailCategorical: React.FC<Props> = ({
