+++ static_report/src/components/shared/Columns/ColumnMatrices/UniquenessMatrix.tsx
@@ -6,10 +6,7 @@ import {
 } from '../../../../utils/formatters';
 import { getColumnMetricRatio } from '../../../../utils/transformers';
 import { MetricCell } from '../ColumnMetrics/MetricCell';
-import {
-  DISTINCTS,
-  DUPLICATES,
-} from '../ColumnCard/ColumnTypeDetail/constants';
+import { DISTINCTS, DUPLICATES } from '../constants';
 
 type Props = { columnDatum: ColumnSchema };
 export function UniquenessMetrics({ columnDatum }: Props) {
