+++ static_report/src/components/shared/Columns/constants.ts
